Move all thinblock related function into thinblock.cpp
