+++ src/thinblock.cpp
@@ -7,6 +7,7 @@
 #include "utiltime.h"
 #include "net.h"
 #include "chainparams.h"
+#include "policy/policy.h"
 #include "pow.h"
 #include "timedata.h"
 #include "main.h"
@@ -18,9 +19,11 @@
 #include <string>
 #include <vector>
 
+using namespace std;
+
 extern CCriticalSection cs_thinblockstats;
 extern CCriticalSection cs_orphancache;
-extern std::map<uint256, COrphanTx> mapOrphanTransactions GUARDED_BY(cs_orphancache);
+extern map<uint256, COrphanTx> mapOrphanTransactions GUARDED_BY(cs_orphancache);
 
 CThinBlock::CThinBlock(const CBlock& block, CBloomFilter& filter)
 {
@@ -49,7 +52,7 @@ CXThinBlock::CXThinBlock(const CBlock& block, CBloomFilter* filter)
 
     unsigned int nTx = block.vtx.size();
     vTxHashes.reserve(nTx);
-    std::set<uint64_t> setPartialTxHash;
+    set<uint64_t> setPartialTxHash;
     for (unsigned int i = 0; i < nTx; i++)
     {
         const uint256 hash256 = block.vtx[i].GetHash();
@@ -76,7 +79,7 @@ CXThinBlock::CXThinBlock(const CBlock& block)
 
     unsigned int nTx = block.vtx.size();
     vTxHashes.reserve(nTx);
-    std::set<uint64_t> setPartialTxHash;
+    set<uint64_t> setPartialTxHash;
 
     LOCK(cs_orphancache);
     for (unsigned int i = 0; i < nTx; i++)
@@ -100,13 +103,13 @@ CXThinBlock::CXThinBlock(const CBlock& block)
     }
 }
 
-CXThinBlockTx::CXThinBlockTx(uint256 blockHash, std::vector<CTransaction>& vTx)
+CXThinBlockTx::CXThinBlockTx(uint256 blockHash, vector<CTransaction>& vTx)
 {
     blockhash = blockHash;
     vMissingTx = vTx;
 }
 
-CXRequestThinBlockTx::CXRequestThinBlockTx(uint256 blockHash, std::set<uint64_t>& setHashesToRequest)
+CXRequestThinBlockTx::CXRequestThinBlockTx(uint256 blockHash, set<uint64_t>& setHashesToRequest)
 {
     blockhash = blockHash;
     setCheapHashesToRequest = setHashesToRequest;
@@ -125,7 +128,7 @@ bool CXThinBlock::CheckBlockHeader(const CBlockHeader& block, CValidationState& 
   return true;
 }
     
-bool CXThinBlock::process(CNode* pfrom, int nSizeThinBlock, std::string strCommand)  // TODO: request from the "best" txn source not necessarily from the block source 
+bool CXThinBlock::process(CNode* pfrom, int nSizeThinBlock, string strCommand)  // TODO: request from the "best" txn source not necessarily from the block source 
 {
     // Xpress Validation - only perform xval if the chaintip matches the last blockhash in the thinblock
     bool fXVal;
@@ -145,7 +148,7 @@ bool CXThinBlock::process(CNode* pfrom, int nSizeThinBlock, std::string strComma
     pfrom->xThinBlockHashes = vTxHashes;
 
     // Create the mapMissingTx from all the supplied tx's in the xthinblock
-    std::map<uint256, CTransaction> mapMissingTx;
+    map<uint256, CTransaction> mapMissingTx;
     BOOST_FOREACH(CTransaction tx, vMissingTx)
       mapMissingTx[tx.GetHash()] = tx;
 
@@ -154,13 +157,13 @@ bool CXThinBlock::process(CNode* pfrom, int nSizeThinBlock, std::string strComma
     int missingCount = 0;
     int unnecessaryCount = 0;
     bool collision = false;
-    std::map<uint64_t, uint256> mapPartialTxHash;
-    std::vector<uint256> memPoolHashes;
+    map<uint64_t, uint256> mapPartialTxHash;
+    vector<uint256> memPoolHashes;
 
     // Do the orphans first before taking the mempool.cs lock, so that we maintain correct locking order.
     {
     LOCK(cs_orphancache);
-    for (std::map<uint256, COrphanTx>::iterator mi = mapOrphanTransactions.begin(); mi != mapOrphanTransactions.end(); ++mi) {
+    for (map<uint256, COrphanTx>::iterator mi = mapOrphanTransactions.begin(); mi != mapOrphanTransactions.end(); ++mi) {
         uint64_t cheapHash = (*mi).first.GetCheapHash();
         if (mapPartialTxHash.count(cheapHash)) //Check for collisions
             collision = true;
@@ -179,7 +182,7 @@ bool CXThinBlock::process(CNode* pfrom, int nSizeThinBlock, std::string strComma
             collision = true;
         mapPartialTxHash[cheapHash] = memPoolHashes[i];
     }
-    for (std::map<uint256, CTransaction>::iterator mi = mapMissingTx.begin(); mi != mapMissingTx.end(); ++mi) {
+    for (map<uint256, CTransaction>::iterator mi = mapMissingTx.begin(); mi != mapMissingTx.end(); ++mi) {
 	uint64_t cheapHash = (*mi).first.GetCheapHash();
         // Check for cheap hash collision. Only mark as collision if the full hash is not the same,
         // because the same tx could have been received into the mempool during the request of the xthinblock.
@@ -196,7 +199,7 @@ bool CXThinBlock::process(CNode* pfrom, int nSizeThinBlock, std::string strComma
     // There is a remote possiblity of a Tx hash collision therefore if it occurs we re-request a normal
     // thinblock which has the full Tx hash data rather than just the truncated hash.
     if (collision) {
-        std::vector<CInv> vGetData;
+        vector<CInv> vGetData;
         vGetData.push_back(CInv(MSG_THINBLOCK, header.GetHash())); 
         pfrom->PushMessage("getdata", vGetData);
         LogPrintf("TX HASH COLLISION for xthinblock: re-requesting a thinblock\n");
@@ -252,7 +255,7 @@ bool CXThinBlock::process(CNode* pfrom, int nSizeThinBlock, std::string strComma
 
         // Update run-time statistics of thin block bandwidth savings
         CThinBlockStats::UpdateInBound(pfrom->nSizeThinBlock, blockSize);
-        std::string ss = CThinBlockStats::ToString();
+        string ss = CThinBlockStats::ToString();
         LogPrint("thin", "thin block stats: %s\n", ss.c_str());
         requester.Received(GetInv(), pfrom, pfrom->nSizeThinBlock);
         HandleBlockMessage(pfrom, strCommand, pfrom->thinBlock,  GetInv());  // clears the thin block
@@ -263,7 +266,7 @@ bool CXThinBlock::process(CNode* pfrom, int nSizeThinBlock, std::string strComma
     else if (pfrom->thinBlockWaitingForTxns > 0) {
         // This marks the end of the transactions we've received. If we get this and we have NOT been able to
         // finish reassembling the block, we need to re-request the transactions we're missing:
-        std::set<uint64_t> setHashesToRequest;
+        set<uint64_t> setHashesToRequest;
         for (size_t i = 0; i < pfrom->thinBlock.vtx.size(); i++) {
 	    if (pfrom->thinBlock.vtx[i].IsNull())
 	        setHashesToRequest.insert(pfrom->xThinBlockHashes[i]);
@@ -287,14 +290,14 @@ void CThinBlockStats::UpdateInBound(uint64_t nThinBlockSize, uint64_t nOriginalB
     CThinBlockStats::nOriginalSize += nOriginalBlockSize;
     CThinBlockStats::nThinSize += nThinBlockSize;
     CThinBlockStats::nBlocks += 1;
-    CThinBlockStats::mapThinBlocksInBound[GetTimeMillis()] = std::pair<uint64_t, uint64_t>(nThinBlockSize, nOriginalBlockSize);
+    CThinBlockStats::mapThinBlocksInBound[GetTimeMillis()] = pair<uint64_t, uint64_t>(nThinBlockSize, nOriginalBlockSize);
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator iter = CThinBlockStats::mapThinBlocksInBound.begin();
+    map<int64_t, pair<uint64_t, uint64_t> >::iterator iter = CThinBlockStats::mapThinBlocksInBound.begin();
     while (iter != CThinBlockStats::mapThinBlocksInBound.end())
     {
-        std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, pair<uint64_t, uint64_t> >::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapThinBlocksInBound.erase(mi);
     }
@@ -307,14 +310,14 @@ void CThinBlockStats::UpdateOutBound(uint64_t nThinBlockSize, uint64_t nOriginal
     CThinBlockStats::nOriginalSize += nOriginalBlockSize;
     CThinBlockStats::nThinSize += nThinBlockSize;
     CThinBlockStats::nBlocks += 1;
-    CThinBlockStats::mapThinBlocksOutBound[GetTimeMillis()] = std::pair<uint64_t, uint64_t>(nThinBlockSize, nOriginalBlockSize);
+    CThinBlockStats::mapThinBlocksOutBound[GetTimeMillis()] = pair<uint64_t, uint64_t>(nThinBlockSize, nOriginalBlockSize);
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator iter = CThinBlockStats::mapThinBlocksOutBound.begin();
+    map<int64_t, pair<uint64_t, uint64_t> >::iterator iter = CThinBlockStats::mapThinBlocksOutBound.begin();
     while (iter != CThinBlockStats::mapThinBlocksOutBound.end())
     {
-        std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, pair<uint64_t, uint64_t> >::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapThinBlocksOutBound.erase(mi);
     }
@@ -329,10 +332,10 @@ void CThinBlockStats::UpdateOutBoundBloomFilter(uint64_t nBloomFilterSize)
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, uint64_t>::iterator iter = CThinBlockStats::mapBloomFiltersOutBound.begin();
+    map<int64_t, uint64_t>::iterator iter = CThinBlockStats::mapBloomFiltersOutBound.begin();
     while (iter != CThinBlockStats::mapBloomFiltersOutBound.end())
     {
-        std::map<int64_t, uint64_t>::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, uint64_t>::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapBloomFiltersOutBound.erase(mi);
     }
@@ -347,10 +350,10 @@ void CThinBlockStats::UpdateInBoundBloomFilter(uint64_t nBloomFilterSize)
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, uint64_t>::iterator iter = CThinBlockStats::mapBloomFiltersInBound.begin();
+    map<int64_t, uint64_t>::iterator iter = CThinBlockStats::mapBloomFiltersInBound.begin();
     while (iter != CThinBlockStats::mapBloomFiltersInBound.end())
     {
-        std::map<int64_t, uint64_t>::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, uint64_t>::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapBloomFiltersInBound.erase(mi);
     }
@@ -366,10 +369,10 @@ void CThinBlockStats::UpdateResponseTime(double nResponseTime)
 
         // Delete any entries that are more than 24 hours old
         int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-        std::map<int64_t, double>::iterator iter = CThinBlockStats::mapThinBlockResponseTime.begin();
+        map<int64_t, double>::iterator iter = CThinBlockStats::mapThinBlockResponseTime.begin();
         while (iter != CThinBlockStats::mapThinBlockResponseTime.end())
         {
-            std::map<int64_t, double>::iterator mi = iter++; // increment to avoid iterator becoming invalid
+            map<int64_t, double>::iterator mi = iter++; // increment to avoid iterator becoming invalid
             if ((*mi).first < nTimeCutoff)
                 CThinBlockStats::mapThinBlockResponseTime.erase(mi);
         }
@@ -386,10 +389,10 @@ void CThinBlockStats::UpdateValidationTime(double nValidationTime)
 
         // Delete any entries that are more than 24 hours old
         int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-        std::map<int64_t, double>::iterator iter = CThinBlockStats::mapThinBlockValidationTime.begin();
+        map<int64_t, double>::iterator iter = CThinBlockStats::mapThinBlockValidationTime.begin();
         while (iter != CThinBlockStats::mapThinBlockValidationTime.end())
         {
-            std::map<int64_t, double>::iterator mi = iter++; // increment to avoid iterator becoming invalid
+            map<int64_t, double>::iterator mi = iter++; // increment to avoid iterator becoming invalid
             if ((*mi).first < nTimeCutoff)
                 CThinBlockStats::mapThinBlockValidationTime.erase(mi);
         }
@@ -405,10 +408,10 @@ void CThinBlockStats::UpdateInBoundReRequestedTx(int nReRequestedTx)
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, int>::iterator iter = CThinBlockStats::mapThinBlocksInBoundReRequestedTx.begin();
+    map<int64_t, int>::iterator iter = CThinBlockStats::mapThinBlocksInBoundReRequestedTx.begin();
     while (iter != CThinBlockStats::mapThinBlocksInBoundReRequestedTx.end())
     {
-        std::map<int64_t, int>::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, int>::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapThinBlocksInBoundReRequestedTx.erase(mi);
     }
@@ -420,7 +423,7 @@ void CThinBlockStats::UpdateMempoolLimiterBytesSaved(unsigned int nBytesSaved)
     CThinBlockStats::nMempoolLimiterBytesSaved += nBytesSaved;
 }
 
-std::string CThinBlockStats::ToString()
+string CThinBlockStats::ToString()
 {
     LOCK(cs_thinblockstats);
 
@@ -432,23 +435,23 @@ std::string CThinBlockStats::ToString()
 	i++;
     }
 
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(2);
+    ostringstream ss;
+    ss << fixed << setprecision(2);
     ss << CThinBlockStats::nBlocks() << " thin " << ((CThinBlockStats::nBlocks() > 1) ? "blocks have" : "block has") << " saved " << size << units[i] << " of bandwidth";
     return ss.str();
 }
 
 // Calculate the xthin percentage compression over the last 24 hours
-std::string CThinBlockStats::InBoundPercentToString()
+string CThinBlockStats::InBoundPercentToString()
 {
     LOCK(cs_thinblockstats);
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator iter = CThinBlockStats::mapThinBlocksInBound.begin();
+    map<int64_t, pair<uint64_t, uint64_t> >::iterator iter = CThinBlockStats::mapThinBlocksInBound.begin();
     while (iter != CThinBlockStats::mapThinBlocksInBound.end())
     {
-        std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, pair<uint64_t, uint64_t> >::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapThinBlocksInBound.erase(mi);
     }
@@ -456,13 +459,13 @@ std::string CThinBlockStats::InBoundPercentToString()
     double nCompressionRate = 0;
     uint64_t nThinSizeTotal = 0;
     uint64_t nOriginalSizeTotal = 0;
-    for (std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator mi = CThinBlockStats::mapThinBlocksInBound.begin(); mi != CThinBlockStats::mapThinBlocksInBound.end(); ++mi) {
+    for (map<int64_t, pair<uint64_t, uint64_t> >::iterator mi = CThinBlockStats::mapThinBlocksInBound.begin(); mi != CThinBlockStats::mapThinBlocksInBound.end(); ++mi) {
         nThinSizeTotal += (*mi).second.first;
         nOriginalSizeTotal += (*mi).second.second;
     }
     // We count up the outbound bloom filters. Outbound bloom filters go with Inbound xthins.
     uint64_t nOutBoundBloomFilterSize = 0;
-    for (std::map<int64_t, uint64_t>::iterator mi = CThinBlockStats::mapBloomFiltersOutBound.begin(); mi != CThinBlockStats::mapBloomFiltersOutBound.end(); ++mi) {
+    for (map<int64_t, uint64_t>::iterator mi = CThinBlockStats::mapBloomFiltersOutBound.begin(); mi != CThinBlockStats::mapBloomFiltersOutBound.end(); ++mi) {
         nOutBoundBloomFilterSize += (*mi).second;
     }
 
@@ -470,23 +473,23 @@ std::string CThinBlockStats::InBoundPercentToString()
     if (nOriginalSizeTotal > 0)
         nCompressionRate = 100 - (100 * (double)(nThinSizeTotal + nOutBoundBloomFilterSize) / nOriginalSizeTotal);
 
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(1);
+    ostringstream ss;
+    ss << fixed << setprecision(1);
     ss << "Compression for " << CThinBlockStats::mapThinBlocksInBound.size() << " Inbound  thinblocks (last 24hrs): " << nCompressionRate << "%";
     return ss.str();
 }
 
 // Calculate the xthin percentage compression over the last 24 hours
-std::string CThinBlockStats::OutBoundPercentToString()
+string CThinBlockStats::OutBoundPercentToString()
 {
     LOCK(cs_thinblockstats);
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator iter = CThinBlockStats::mapThinBlocksOutBound.begin();
+    map<int64_t, pair<uint64_t, uint64_t> >::iterator iter = CThinBlockStats::mapThinBlocksOutBound.begin();
     while (iter != CThinBlockStats::mapThinBlocksOutBound.end())
     {
-        std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, pair<uint64_t, uint64_t> >::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapThinBlocksOutBound.erase(mi);
     }
@@ -494,36 +497,36 @@ std::string CThinBlockStats::OutBoundPercentToString()
     double nCompressionRate = 0;
     uint64_t nThinSizeTotal = 0;
     uint64_t nOriginalSizeTotal = 0;
-    for (std::map<int64_t, std::pair<uint64_t, uint64_t> >::iterator mi = CThinBlockStats::mapThinBlocksOutBound.begin(); mi != CThinBlockStats::mapThinBlocksOutBound.end(); ++mi) {
+    for (map<int64_t, pair<uint64_t, uint64_t> >::iterator mi = CThinBlockStats::mapThinBlocksOutBound.begin(); mi != CThinBlockStats::mapThinBlocksOutBound.end(); ++mi) {
         nThinSizeTotal += (*mi).second.first;
         nOriginalSizeTotal += (*mi).second.second;
     }
     // We count up the inbound bloom filters. Inbound bloom filters go with Outbound xthins.
     uint64_t nInBoundBloomFilterSize = 0;
-    for (std::map<int64_t, uint64_t>::iterator mi = CThinBlockStats::mapBloomFiltersInBound.begin(); mi != CThinBlockStats::mapBloomFiltersInBound.end(); ++mi) {
+    for (map<int64_t, uint64_t>::iterator mi = CThinBlockStats::mapBloomFiltersInBound.begin(); mi != CThinBlockStats::mapBloomFiltersInBound.end(); ++mi) {
         nInBoundBloomFilterSize += (*mi).second;
     }
 
     if (nOriginalSizeTotal > 0)
         nCompressionRate = 100 - (100 * (double)(nThinSizeTotal + nInBoundBloomFilterSize) / nOriginalSizeTotal);
 
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(1);
+    ostringstream ss;
+    ss << fixed << setprecision(1);
     ss << "Compression for " << CThinBlockStats::mapThinBlocksOutBound.size() << " Outbound thinblocks (last 24hrs): " << nCompressionRate << "%";
     return ss.str();
 }
 
 // Calculate the average inbound xthin bloom filter size
-std::string CThinBlockStats::InBoundBloomFiltersToString()
+string CThinBlockStats::InBoundBloomFiltersToString()
 {
     LOCK(cs_thinblockstats);
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, uint64_t>::iterator iter = CThinBlockStats::mapBloomFiltersInBound.begin();
+    map<int64_t, uint64_t>::iterator iter = CThinBlockStats::mapBloomFiltersInBound.begin();
     while (iter != CThinBlockStats::mapBloomFiltersInBound.end())
     {
-        std::map<int64_t, uint64_t>::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, uint64_t>::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapBloomFiltersInBound.erase(mi);
     }
@@ -531,7 +534,7 @@ std::string CThinBlockStats::InBoundBloomFiltersToString()
     uint64_t nInBoundBloomFilters = 0;
     uint64_t nInBoundBloomFilterSize = 0;
     double avgBloomSize = 0;
-    for (std::map<int64_t, uint64_t>::iterator mi = CThinBlockStats::mapBloomFiltersInBound.begin(); mi != CThinBlockStats::mapBloomFiltersInBound.end(); ++mi) {
+    for (map<int64_t, uint64_t>::iterator mi = CThinBlockStats::mapBloomFiltersInBound.begin(); mi != CThinBlockStats::mapBloomFiltersInBound.end(); ++mi) {
         nInBoundBloomFilterSize += (*mi).second;
         nInBoundBloomFilters += 1;
     }
@@ -545,23 +548,23 @@ std::string CThinBlockStats::InBoundBloomFiltersToString()
 	i++;
     }
 
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(1);
+    ostringstream ss;
+    ss << fixed << setprecision(1);
     ss << "Inbound bloom filter size (last 24hrs) AVG: " << avgBloomSize << units[i];
     return ss.str();
 }
 
 // Calculate the average inbound xthin bloom filter size
-std::string CThinBlockStats::OutBoundBloomFiltersToString()
+string CThinBlockStats::OutBoundBloomFiltersToString()
 {
     LOCK(cs_thinblockstats);
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, uint64_t>::iterator iter = CThinBlockStats::mapBloomFiltersOutBound.begin();
+    map<int64_t, uint64_t>::iterator iter = CThinBlockStats::mapBloomFiltersOutBound.begin();
     while (iter != CThinBlockStats::mapBloomFiltersOutBound.end())
     {
-        std::map<int64_t, uint64_t>::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, uint64_t>::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapBloomFiltersOutBound.erase(mi);
     }
@@ -569,7 +572,7 @@ std::string CThinBlockStats::OutBoundBloomFiltersToString()
     uint64_t nOutBoundBloomFilters = 0;
     uint64_t nOutBoundBloomFilterSize = 0;
     double avgBloomSize = 0;
-    for (std::map<int64_t, uint64_t>::iterator mi = CThinBlockStats::mapBloomFiltersOutBound.begin(); mi != CThinBlockStats::mapBloomFiltersOutBound.end(); ++mi) {
+    for (map<int64_t, uint64_t>::iterator mi = CThinBlockStats::mapBloomFiltersOutBound.begin(); mi != CThinBlockStats::mapBloomFiltersOutBound.end(); ++mi) {
         nOutBoundBloomFilterSize += (*mi).second;
         nOutBoundBloomFilters += 1;
     }
@@ -583,23 +586,23 @@ std::string CThinBlockStats::OutBoundBloomFiltersToString()
 	i++;
     }
 
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(1);
+    ostringstream ss;
+    ss << fixed << setprecision(1);
     ss << "Outbound bloom filter size (last 24hrs) AVG: " << avgBloomSize << units[i];
     return ss.str();
 }
 // Calculate the xthin percentage compression over the last 24 hours
-std::string CThinBlockStats::ResponseTimeToString()
+string CThinBlockStats::ResponseTimeToString()
 {
     LOCK(cs_thinblockstats);
 
-    std::vector<double> vResponseTime;
+    vector<double> vResponseTime;
 
     double nResponseTimeAverage = 0;
     double nPercentile = 0;
     double nTotalResponseTime = 0;
     double nTotalEntries = 0;
-    for (std::map<int64_t, double>::iterator mi = CThinBlockStats::mapThinBlockResponseTime.begin(); mi != CThinBlockStats::mapThinBlockResponseTime.end(); ++mi) {
+    for (map<int64_t, double>::iterator mi = CThinBlockStats::mapThinBlockResponseTime.begin(); mi != CThinBlockStats::mapThinBlockResponseTime.end(); ++mi) {
         nTotalEntries += 1;
         nTotalResponseTime += (*mi).second;
         vResponseTime.push_back((*mi).second);
@@ -610,28 +613,28 @@ std::string CThinBlockStats::ResponseTimeToString()
 
         // Calculate the 95th percentile
         uint64_t nPercentileElement = static_cast<int>((nTotalEntries * 0.95) + 0.5) - 1;
-        std::sort(vResponseTime.begin(), vResponseTime.end());
+        sort(vResponseTime.begin(), vResponseTime.end());
         nPercentile = vResponseTime[nPercentileElement];
     }
 
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(2);
+    ostringstream ss;
+    ss << fixed << setprecision(2);
     ss << "Response time   (last 24hrs) AVG:" << nResponseTimeAverage << ", 95th pcntl:" << nPercentile;
     return ss.str();
 }
 
 // Calculate the xthin percentage compression over the last 24 hours
-std::string CThinBlockStats::ValidationTimeToString()
+string CThinBlockStats::ValidationTimeToString()
 {
     LOCK(cs_thinblockstats);
 
-    std::vector<double> vValidationTime;
+    vector<double> vValidationTime;
 
     double nValidationTimeAverage = 0;
     double nPercentile = 0;
     double nTotalValidationTime = 0;
     double nTotalEntries = 0;
-    for (std::map<int64_t, double>::iterator mi = CThinBlockStats::mapThinBlockValidationTime.begin(); mi != CThinBlockStats::mapThinBlockValidationTime.end(); ++mi) {
+    for (map<int64_t, double>::iterator mi = CThinBlockStats::mapThinBlockValidationTime.begin(); mi != CThinBlockStats::mapThinBlockValidationTime.end(); ++mi) {
         nTotalEntries += 1;
         nTotalValidationTime += (*mi).second;
         vValidationTime.push_back((*mi).second);
@@ -642,27 +645,27 @@ std::string CThinBlockStats::ValidationTimeToString()
 
         // Calculate the 95th percentile
         uint64_t nPercentileElement = static_cast<int>((nTotalEntries * 0.95) + 0.5) - 1;
-        std::sort(vValidationTime.begin(), vValidationTime.end());
+        sort(vValidationTime.begin(), vValidationTime.end());
         nPercentile = vValidationTime[nPercentileElement];
     }
 
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(2);
+    ostringstream ss;
+    ss << fixed << setprecision(2);
     ss << "Validation time (last 24hrs) AVG:" << nValidationTimeAverage << ", 95th pcntl:" << nPercentile;
     return ss.str();
 }
 
 // Calculate the xthin percentage compression over the last 24 hours
-std::string CThinBlockStats::ReRequestedTxToString()
+string CThinBlockStats::ReRequestedTxToString()
 {
     LOCK(cs_thinblockstats);
 
     // Delete any entries that are more than 24 hours old
     int64_t nTimeCutoff = GetTimeMillis() - 60*60*24*1000;
-    std::map<int64_t, int>::iterator iter = CThinBlockStats::mapThinBlocksInBoundReRequestedTx.begin();
+    map<int64_t, int>::iterator iter = CThinBlockStats::mapThinBlocksInBoundReRequestedTx.begin();
     while (iter != CThinBlockStats::mapThinBlocksInBoundReRequestedTx.end())
     {
-        std::map<int64_t, int>::iterator mi = iter++; // increment to avoid iterator becoming invalid
+        map<int64_t, int>::iterator mi = iter++; // increment to avoid iterator becoming invalid
         if ((*mi).first < nTimeCutoff)
             CThinBlockStats::mapThinBlocksInBoundReRequestedTx.erase(mi);
     }
@@ -670,7 +673,7 @@ std::string CThinBlockStats::ReRequestedTxToString()
     double nReRequestRate = 0;
     uint64_t nTotalReRequests = 0;
     uint64_t nTotalReRequestedTxs = 0;
-    for (std::map<int64_t, int>::iterator mi = CThinBlockStats::mapThinBlocksInBoundReRequestedTx.begin(); mi != CThinBlockStats::mapThinBlocksInBoundReRequestedTx.end(); ++mi) {
+    for (map<int64_t, int>::iterator mi = CThinBlockStats::mapThinBlocksInBoundReRequestedTx.begin(); mi != CThinBlockStats::mapThinBlocksInBoundReRequestedTx.end(); ++mi) {
         nTotalReRequests += 1;
         nTotalReRequestedTxs += (*mi).second;
     }
@@ -678,13 +681,13 @@ std::string CThinBlockStats::ReRequestedTxToString()
     if ( CThinBlockStats::mapThinBlocksInBound.size() > 0)
         nReRequestRate = 100 * (double)nTotalReRequests / CThinBlockStats::mapThinBlocksInBound.size();
 
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(1);
+    ostringstream ss;
+    ss << fixed << setprecision(1);
     ss << "Tx re-request rate (last 24hrs): " << nReRequestRate << "% Total re-requests:" << nTotalReRequests;
     return ss.str();
 }
 
-std::string CThinBlockStats::MempoolLimiterBytesSavedToString()
+string CThinBlockStats::MempoolLimiterBytesSavedToString()
 {
     LOCK(cs_thinblockstats);
 
@@ -696,8 +699,410 @@ std::string CThinBlockStats::MempoolLimiterBytesSavedToString()
 	i++;
     }
 
-    std::ostringstream ss;
-    ss << std::fixed << std::setprecision(2);
+    ostringstream ss;
+    ss << fixed << setprecision(2);
     ss << "Thinblock mempool limiting has saved " << size << units[i] << " of bandwidth";
     return ss.str();
 }
+
+
+
+bool HaveConnectThinblockNodes()
+{
+    // Strip the port from then list of all the current in and outbound ip addresses
+    vector<string> vNodesIP;
+    {
+        LOCK(cs_vNodes);
+        BOOST_FOREACH (CNode* pnode, vNodes) {
+           int pos = pnode->addrName.rfind(":");
+           if (pos <= 0 )
+               vNodesIP.push_back(pnode->addrName);
+           else
+               vNodesIP.push_back(pnode->addrName.substr(0, pos));
+        }
+    }
+
+    // Create a set used to check for cross connected nodes.
+    // A cross connected node is one where we have a connect-thinblock connection to
+    // but we also have another inbound connection which is also using
+    // connect-thinblock. In those cases we have created a dead-lock where no blocks 
+    // can be downloaded unless we also have at least one additional connect-thinblock 
+    // connection to a different node.
+    set<string> nNotCrossConnected;
+ 
+    int nConnectionsOpen = 0;
+    BOOST_FOREACH(const string& strAddrNode, mapMultiArgs["-connect-thinblock"]) {
+        string strThinblockNode;
+        int pos = strAddrNode.rfind(":");
+        if (pos <= 0 )
+            strThinblockNode = strAddrNode;
+        else
+            strThinblockNode = strAddrNode.substr(0, pos);
+        BOOST_FOREACH(string strAddr, vNodesIP) {
+            if (strAddr == strThinblockNode) {
+                nConnectionsOpen++;
+                if (!nNotCrossConnected.count(strAddr))
+                    nNotCrossConnected.insert(strAddr);
+                else
+                    nNotCrossConnected.erase(strAddr);
+            }
+        }
+    }
+    if (nNotCrossConnected.size() > 0)
+        return true;
+    else if (nConnectionsOpen > 0)
+        LogPrint("thin", "You have a cross connected thinblock node - we may download regular blocks until you resolve the issue\n");
+    return false; // Connections are either not open or they are cross connected.
+} 
+
+
+bool HaveThinblockNodes()
+{
+    {
+        LOCK(cs_vNodes);
+        BOOST_FOREACH (CNode* pnode, vNodes)
+            if (pnode->ThinBlockCapable())
+                return true;
+    }
+    return false;
+}
+
+bool CheckThinblockTimer(uint256 hash)
+{
+    if (!mapThinBlockTimer.count(hash)) {
+        mapThinBlockTimer[hash] = GetTimeMillis();
+        LogPrint("thin", "Starting Preferential Thinblock timer\n");
+    }
+    else {
+        // Check that we have not exceeded the 10 second limit.
+        // If we have then we want to return false so that we can
+        // proceed to download a regular block instead.
+        uint64_t elapsed = GetTimeMillis() - mapThinBlockTimer[hash];
+        if (elapsed > 10000) {
+            LogPrint("thin", "Preferential Thinblock timer exceeded - downloading regular block instead\n");
+            return false;
+        }
+    }
+    return true;
+}
+
+void ClearThinBlockTimer(uint256 hash)
+{
+    if (mapThinBlockTimer.count(hash)) {
+        mapThinBlockTimer.erase(hash);
+        LogPrint("thin", "Clearing Preferential Thinblock timer\n");
+    }
+}
+
+bool IsThinBlocksEnabled() 
+{
+    return GetBoolArg("-use-thinblocks", true);
+}
+
+bool CanThinBlockBeDownloaded(CNode* pto)
+{
+    if (pto->ThinBlockCapable() && !GetBoolArg("-connect-thinblock-force", false))
+        return true;
+    else if (pto->ThinBlockCapable() && GetBoolArg("-connect-thinblock-force", false)) {
+        // If connect-thinblock-force is true then we have to check that this node is in fact a connect-thinblock node.
+
+        // When -connect-thinblock-force is true we will only download thinblocks from a peer or peers that
+        // are using -connect-thinblock=<ip>.  This is an undocumented setting used for setting up performance testing
+        // of thinblocks, such as, going over the GFC and needing to have thinblocks always come from the same peer or 
+        // group of peers.  Also, this is a one way street.  Thinblocks will flow ONLY from the remote peer to the peer
+        // that has invoked -connect-thinblock.
+
+        // Check if this node is also a connect-thinblock node
+        BOOST_FOREACH(const string& strAddrNode, mapMultiArgs["-connect-thinblock"])
+            if (pto->addrName == strAddrNode)
+                return true;
+    }
+    return false;
+}
+
+void ConnectToThinBlockNodes()
+{
+    // Connect to specific addresses
+    if (mapArgs.count("-connect-thinblock") && mapMultiArgs["-connect-thinblock"].size() > 0)
+    {
+        BOOST_FOREACH(const string& strAddr, mapMultiArgs["-connect-thinblock"])
+        {
+            CAddress addr;
+            //NOTE: Because the only nodes we are connecting to here are the ones the user put in their
+            //      bitcoin.conf/commandline args as "-connect-thinblock", we don't use the semaphore to limit outbound connections
+            OpenNetworkConnection(addr, NULL, strAddr.c_str());
+            MilliSleep(500);
+        }
+    }
+}
+
+void CheckNodeSupportForThinBlocks()
+{
+    if(IsThinBlocksEnabled()) {
+        //BU: Enforce cs_vNodes lock held external to FindNode function calls to prevent use-after-free errors
+        LOCK(cs_vNodes);
+        // Check that a nodes pointed to with connect-thinblock actually supports thinblocks
+        BOOST_FOREACH(string& strAddr, mapMultiArgs["-connect-thinblock"]) {
+            if(CNode* pnode = FindNode(strAddr)) {
+                if(!pnode->ThinBlockCapable()) {
+                    LogPrintf("ERROR: You are trying to use connect-thinblocks but to a node that does not support it - Protocol Version: %d peer=%d\n", 
+                               pnode->nVersion, pnode->id);
+                }
+            }
+        }
+    }
+}
+
+void SendXThinBlock(CBlock &block, CNode* pfrom, const CInv &inv)
+{
+    if (inv.type == MSG_XTHINBLOCK)
+    {
+        CXThinBlock xThinBlock(block, pfrom->pThinBlockFilter);
+        int nSizeBlock = ::GetSerializeSize(block, SER_NETWORK, PROTOCOL_VERSION);
+        if (xThinBlock.collision == true) // If there is a cheapHash collision in this block then send a normal thinblock
+        {
+            CThinBlock thinBlock(block, *pfrom->pThinBlockFilter);
+            int nSizeThinBlock = ::GetSerializeSize(xThinBlock, SER_NETWORK, PROTOCOL_VERSION);
+            if (nSizeThinBlock < nSizeBlock) {
+                pfrom->PushMessage(NetMsgType::THINBLOCK, thinBlock);
+                CThinBlockStats::UpdateOutBound(nSizeThinBlock, nSizeBlock);
+                LogPrint("thin", "TX HASH COLLISION: Sent thinblock - size: %d vs block size: %d => tx hashes: %d transactions: %d  peer: %s (%d)\n", nSizeThinBlock, nSizeBlock, xThinBlock.vTxHashes.size(), xThinBlock.vMissingTx.size(), pfrom->addrName.c_str(), pfrom->id);
+            }
+            else {
+                pfrom->PushMessage(NetMsgType::BLOCK, block);
+                LogPrint("thin", "Sent regular block instead - xthinblock size: %d vs block size: %d => tx hashes: %d transactions: %d  peer: %s (%d)\n", nSizeThinBlock, nSizeBlock, xThinBlock.vTxHashes.size(), xThinBlock.vMissingTx.size(), pfrom->addrName.c_str(), pfrom->id);
+            }
+        }
+        else // Send an xThinblock
+        {
+            // Only send a thinblock if smaller than a regular block
+            int nSizeThinBlock = ::GetSerializeSize(xThinBlock, SER_NETWORK, PROTOCOL_VERSION);
+            if (nSizeThinBlock < nSizeBlock) {
+                CThinBlockStats::UpdateOutBound(nSizeThinBlock, nSizeBlock);
+                pfrom->PushMessage(NetMsgType::XTHINBLOCK, xThinBlock);
+                LogPrint("thin", "Sent xthinblock - size: %d vs block size: %d => tx hashes: %d transactions: %d peer: %s (%d)\n", nSizeThinBlock, nSizeBlock, xThinBlock.vTxHashes.size(), xThinBlock.vMissingTx.size(), pfrom->addrName.c_str(), pfrom->id);
+            }
+            else {
+                pfrom->PushMessage(NetMsgType::BLOCK, block);
+                LogPrint("thin", "Sent regular block instead - xthinblock size: %d vs block size: %d => tx hashes: %d transactions: %d  peer: %s (%d)\n", nSizeThinBlock, nSizeBlock, xThinBlock.vTxHashes.size(), xThinBlock.vMissingTx.size(), pfrom->addrName.c_str(), pfrom->id);
+            }
+        }
+    }
+    else if (inv.type == MSG_THINBLOCK)
+    {
+        CThinBlock thinBlock(block, *pfrom->pThinBlockFilter);
+        int nSizeBlock = ::GetSerializeSize(block, SER_NETWORK, PROTOCOL_VERSION);
+        int nSizeThinBlock = ::GetSerializeSize(thinBlock, SER_NETWORK, PROTOCOL_VERSION);
+        if (nSizeThinBlock < nSizeBlock) { // Only send a thinblock if smaller than a regular block
+            CThinBlockStats::UpdateOutBound(nSizeThinBlock, nSizeBlock);
+            pfrom->PushMessage(NetMsgType::THINBLOCK, thinBlock);
+            LogPrint("thin", "Sent thinblock - size: %d vs block size: %d => tx hashes: %d transactions: %d  peer: %s (%d)\n", nSizeThinBlock, nSizeBlock, thinBlock.vTxHashes.size(), thinBlock.vMissingTx.size(), pfrom->addrName.c_str(), pfrom->id);
+        }
+        else {
+            pfrom->PushMessage(NetMsgType::BLOCK, block);
+            LogPrint("thin", "Sent regular block instead - thinblock size: %d vs block size: %d => tx hashes: %d transactions: %d  peer: %s (%d)\n", nSizeThinBlock, nSizeBlock, thinBlock.vTxHashes.size(), thinBlock.vMissingTx.size(), pfrom->addrName.c_str(), pfrom->id);
+        }
+    }
+    else
+      {
+	assert(0);  // inv type is not correct 
+      }
+    pfrom->blocksSent += 1;
+}
+
+void BuildSeededBloomFilter(CBloomFilter& filterMemPool, vector<uint256>& vOrphanHashes, uint256 hash)
+{
+    int64_t nStartTimer = GetTimeMillis();
+    seed_insecure_rand();
+    set<uint256> setHighScoreMemPoolHashes;
+    set<uint256> setPriorityMemPoolHashes;
+
+    // How much of the block should be dedicated to high-priority transactions.
+    // Logically this should be the same size as the DEFAULT_BLOCK_PRIORITY_SIZE however,
+    // we can't be sure that a miner won't decide to mine more high priority txs and therefore
+    // by including a full blocks worth of high priority tx's we cover every scenario.  And when we
+    // go on to add the high fee tx's there will be an intersection between the two which then makes 
+    // the total number of tx's that go into the bloom filter smaller than just the sum of the two.  
+    uint64_t nBlockPrioritySize = LargestBlockSeen() * 1.5;
+
+    // Largest projected block size used to add the high fee transactions.  We multiply it by an
+    // additional factor to take into account that miners may have slighty different policies when selecting
+    // high fee tx's from the pool.
+    uint64_t nBlockMaxProjectedSize = LargestBlockSeen() * 1.5;
+
+    vector<TxCoinAgePriority> vPriority;
+    TxCoinAgePriorityCompare pricomparer;
+    {
+        LOCK2(cs_main, mempool.cs);
+        if (mempool.mapTx.size() > 0) 
+        {
+            CBlockIndex* pindexPrev = chainActive.Tip();
+            const int nHeight = pindexPrev->nHeight + 1;
+            const int64_t nMedianTimePast = pindexPrev->GetMedianTimePast();
+
+            int64_t nLockTimeCutoff = (STANDARD_LOCKTIME_VERIFY_FLAGS & LOCKTIME_MEDIAN_TIME_PAST)
+                                    ? nMedianTimePast
+                                    : GetAdjustedTime();
+
+            // Create a sorted list of transactions and their updated priorities.  This will be used to fill
+            // the mempoolhashes with the expected priority area of the next block.  We will multiply this by
+            // a factor of ? to account for any differences between the "Miners".
+            vPriority.reserve(mempool.mapTx.size());
+            for (CTxMemPool::indexed_transaction_set::iterator mi = mempool.mapTx.begin();
+                 mi != mempool.mapTx.end(); mi++)
+            {
+                double dPriority = mi->GetPriority(nHeight);
+                CAmount dummy;
+                mempool.ApplyDeltas(mi->GetTx().GetHash(), dPriority, dummy);
+                vPriority.push_back(TxCoinAgePriority(dPriority, mi));
+            }
+            make_heap(vPriority.begin(), vPriority.end(), pricomparer);
+
+            uint64_t nPrioritySize = 0;
+            CTxMemPool::txiter iter;
+            for (uint64_t i = 0; i < vPriority.size(); i++)
+            {
+                nPrioritySize += vPriority[i].second->GetTxSize();
+                if (nPrioritySize > nBlockPrioritySize)
+                    break;
+                setPriorityMemPoolHashes.insert(vPriority[i].second->GetTx().GetHash());
+
+                // Add children.  We don't need to look for parents here since they will all be parents.
+                iter = mempool.mapTx.project<0>(vPriority[i].second);
+                BOOST_FOREACH(CTxMemPool::txiter child, mempool.GetMemPoolChildren(iter))
+                {
+                    uint256 childHash = child->GetTx().GetHash();
+                    if (!setPriorityMemPoolHashes.count(childHash)) {
+                        setPriorityMemPoolHashes.insert(childHash);
+                        nPrioritySize += child->GetTxSize();
+                        LogPrint("bloom", "add priority child %s with fee %d modified fee %d size %d clearatentry %d priority %f\n", 
+                                       child->GetTx().GetHash().ToString(), child->GetFee(), child->GetModifiedFee(), 
+                                       child->GetTxSize(), child->WasClearAtEntry(), child->GetPriority(nHeight));
+                    }
+                }            
+            }
+
+            // Create a list of high score transactions. We will multiply this by
+            // a factor of ? to account for any differences between the way Miners include tx's
+            CTxMemPool::indexed_transaction_set::nth_index<3>::type::iterator mi = mempool.mapTx.get<3>().begin();
+            uint64_t nBlockSize = 0;
+            while (mi != mempool.mapTx.get<3>().end())
+            {
+                CTransaction tx = mi->GetTx();
+
+                if (!IsFinalTx(tx, nHeight, nLockTimeCutoff)) {
+                    LogPrint("bloom", "tx %s is not final\n", tx.GetHash().ToString());
+                    mi++;
+                    continue;
+                }
+
+                // If this tx is not accounted for already in the priority set then continue and add
+                // it to the high score set if it can be and also add any parents or children.  Also add
+                // children and parents to the priority set tx's if they have any.
+                iter = mempool.mapTx.project<0>(mi);
+                if (!setHighScoreMemPoolHashes.count(tx.GetHash()))
+                {
+                    LogPrint("bloom", "next tx is %s blocksize %d fee %d modified fee %d size %d clearatentry %d priority %f\n", 
+                                   mi->GetTx().GetHash().ToString(), nBlockSize, mi->GetFee(), mi->GetModifiedFee(), mi->GetTxSize(), 
+                                   mi->WasClearAtEntry(), mi->GetPriority(nHeight));
+
+                    // add tx to the set: we don't know if this is a parent or child yet.
+                    setHighScoreMemPoolHashes.insert(tx.GetHash());
+
+                    // Add any parent tx's
+                    bool fChild = false;
+                    BOOST_FOREACH(CTxMemPool::txiter parent, mempool.GetMemPoolParents(iter))
+                    {
+                        fChild = true;
+                        uint256 parentHash = parent->GetTx().GetHash();
+                        if (!setHighScoreMemPoolHashes.count(parentHash)) {
+                            setHighScoreMemPoolHashes.insert(parentHash);
+                            LogPrint("bloom", "add high score parent %s with blocksize %d fee %d modified fee %d size %d clearatentry %d priority %f\n", 
+                                           parent->GetTx().GetHash().ToString(), nBlockSize, parent->GetFee(), parent->GetModifiedFee(), 
+                                           parent->GetTxSize(), parent->WasClearAtEntry(), parent->GetPriority(nHeight));
+                        }
+                    }
+
+                    // Now add any children tx's.
+                    bool fHasChildren = false;
+                    BOOST_FOREACH(CTxMemPool::txiter child, mempool.GetMemPoolChildren(iter))
+                    {
+                        fHasChildren = true;
+                        uint256 childHash = child->GetTx().GetHash();
+                        if (!setHighScoreMemPoolHashes.count(childHash)) {
+                            setHighScoreMemPoolHashes.insert(childHash);
+                            LogPrint("bloom", "add high score child %s with blocksize %d fee %d modified fee %d size %d clearatentry %d priority %f\n", 
+                                           child->GetTx().GetHash().ToString(), nBlockSize, child->GetFee(), child->GetModifiedFee(), 
+                                           child->GetTxSize(), child->WasClearAtEntry(), child->GetPriority(nHeight));
+                        }
+                    }
+
+                    // If a tx with no parents and no children, then we increment this block size.  
+                    // We don't want to add parents and children to the size because for tx's with many children, miners may not mine them
+                    // as they are not as profitable but we still have to add their hash to the bloom filter in case they do.
+                    if (!fChild && !fHasChildren)
+                        nBlockSize += mi->GetTxSize();
+                }
+
+                if (nBlockSize >  nBlockMaxProjectedSize)
+                    break;
+
+                mi++;
+            }
+        }
+    }
+    LogPrint("thin", "Bloom Filter Targeting completed in:%d (ms)\n", GetTimeMillis() - nStartTimer);
+    nStartTimer= GetTimeMillis(); // reset the timer
+
+    // We set the beginning of our growth algortithm to the time we request our first xthin.  We do this here
+    // rather than setting up a global variable in init.cpp.  This has more to do with potential merge conflicts
+    // with BU than any other technical reason.
+    static int64_t nStartGrowth = GetTime();
+
+    // Tuning knobs for the false positive growth algorithm
+    static uint8_t nHoursToGrow = 72; // number of hours until maximum growth for false positive rate
+    //static double nGrowthCoefficient = 0.7676; // use for nMinFalsePositive = 0.0001 and nMaxFalsePositive = 0.01 for 6 hour growth period
+    //static double nGrowthCoefficient = 0.8831; // use for nMinFalsePositive = 0.0001 and nMaxFalsePositive = 0.02 for 6 hour growth period
+    //static double nGrowthCoefficient = 0.1921; // use for nMinFalsePositive = 0.0001 and nMaxFalsePositive = 0.01 for 24 hour growth period
+    static double nGrowthCoefficient = 0.0544; // use for nMinFalsePositive = 0.0001 and nMaxFalsePositive = 0.005 for 72 hour growth period
+    static double nMinFalsePositive = 0.0001; // starting value for false positive 
+    static double nMaxFalsePositive = 0.005; // maximum false positive rate at end of decay
+    // TODO: automatically calculate the nGrowthCoefficient from nHoursToGrow, nMinFalsePositve and nMaxFalsePositive
+
+    // Count up all the transactions that we'll be putting into the filter, removing any duplicates
+    BOOST_FOREACH(uint256 txHash, setHighScoreMemPoolHashes)
+        if (setPriorityMemPoolHashes.count(txHash))
+            setPriorityMemPoolHashes.erase(txHash);
+
+    unsigned int nSelectedTxHashes = setHighScoreMemPoolHashes.size() + vOrphanHashes.size() + setPriorityMemPoolHashes.size();
+    unsigned int nElements = max(nSelectedTxHashes, (unsigned int)1); // Must make sure nElements is greater than zero or will assert
+
+    // Calculate the new False Positive rate.  
+    // We increase the false positive rate as time increases, starting at nMinFalsePositive and with growth governed by nGrowthCoefficient,
+    // using the simple exponential growth function as follows:
+    // y = (starting or minimum fprate: nMinFalsePositive) * e ^ (time in hours from start * nGrowthCoefficient)
+    int64_t nTimePassed = GetTime() - nStartGrowth;
+    double nFPRate = nMinFalsePositive * exp (((double)(nTimePassed) / 3600) * nGrowthCoefficient);
+    if (nTimePassed > nHoursToGrow * 3600)
+        nFPRate = nMaxFalsePositive;
+
+    filterMemPool = CBloomFilter(nElements, nFPRate, insecure_rand(), BLOOM_UPDATE_ALL);
+    LogPrint("thin", "FPrate: %f Num elements in bloom filter:%d high priority txs:%d high fee txs:%d orphans:%d total txs in mempool:%d\n", 
+              nFPRate, nElements, setPriorityMemPoolHashes.size(), 
+              setHighScoreMemPoolHashes.size(), vOrphanHashes.size(), mempool.mapTx.size());
+
+    // Add the selected tx hashes to the bloom filter
+    BOOST_FOREACH(uint256 txHash, setPriorityMemPoolHashes)
+        filterMemPool.insert(txHash);
+    BOOST_FOREACH(uint256 txHash, setHighScoreMemPoolHashes)
+        filterMemPool.insert(txHash);
+    BOOST_FOREACH(uint256 txHash, vOrphanHashes)
+        filterMemPool.insert(txHash);
+    uint64_t nSizeFilter = ::GetSerializeSize(filterMemPool, SER_NETWORK, PROTOCOL_VERSION);
+    LogPrint("thin", "Created bloom filter: %d bytes for block: %s in:%d (ms)\n", nSizeFilter, hash.ToString(), GetTimeMillis() - nStartTimer);
+    CThinBlockStats::UpdateOutBoundBloomFilter(nSizeFilter);
+}
+
+
+
+
